add separate fusion profile hook
