+++ openpype/hosts/fusion/__init__.py
@@ -1,10 +1,12 @@
 from .addon import (
     FusionAddon,
     FUSION_HOST_DIR,
+    FUSION_PROFILE_VERSION
 )
 
 
 __all__ = (
     "FusionAddon",
     "FUSION_HOST_DIR",
+    "FUSION_PROFILE_VERSION"
 )
